Parallelise hiccup-based-parser
Parsing appears to be the costliest step of any analysis, with most time
being spent in Instaparse. In order to make it faster this commit makes
parsing parallel.

This patch moves parse-fn to a separate map call in parse-from. This
allows parallel parsing with pmap.

Tested on a 4 core CPU, this reduces the time spent on a summary analysis
of a 3800 commit git2 history from 1.650s to 0.996s.