Finer parallelisation of parsing with reducers
This commit loads the sequence generated by as-entry-tokens into a vector.
The vector is folded using clojure.core.reducers/fold, which splits it into
subvectors of 32 entries and parses in parallel.

Tested on a 4 core CPU, this reduces the time spent on a summary
analysis of a 3800 commit git2 history from 0.996s to 0.925s.
Faster processing is noticeable also for shorter inputs.

The cost this change introduces is eagerly preloading the entire collection
returned by as-entry-tokens, just like before 8a2cb3510c.